Lesson 6

+++ api/src/main/java/com/paltvlad/market/api/CartDto.java
@@ -1,6 +1,7 @@
 package com.paltvlad.market.api;
 
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -11,7 +12,7 @@ public class CartDto {
 
     private List<CartItemDto> items;
 
-    private double totalPrice;
+    private BigDecimal totalPrice;
 
     public List<CartItemDto> getItems() {
         return items;
@@ -21,11 +22,11 @@ public class CartDto {
         this.items = items;
     }
 
-    public double getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(double totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
 }
